Library re-organization to allow for easy porting to different platforms.  No changes have been brought to the API or functionality, and all the examples are left completely untouched and compile/run fine.  The main purpose was to segregate the deeply Arduino/AVR-specific details in their own area, as this will make implementing the functionality on various platforms (e.g. those without support for PROGMEM) much easier.
In addition, a number of small modifications were made to make the source lighter (e.g. replacing all the ADAFRUIT_FONA_DEBUG #ifdefs with a set of compile-time selected macros) and reduce the memory load by maximizing use of flash strings (not by much, but 4% on a Duemilanove can make a difference).