Switch from Arduino to mraa
Everything that was arduino related got replaced or is commented out.
A lot of C style was replaced by C++.
All changes made are suited for the FONA3G and intel edison platform.
Since it uses the mraa library it might also work with the raspberry pi,
beagle bone black ...